requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
diff-match-patch==20241021
Django==5.2.3
django-import-export==4.3.7
djangorestframework==3.16.0
idna==3.10
import-export==0.3.1
requests==2.32.4
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
urllib3==2.4.0
